+++ scripts/os/ubuntu/profile-ubuntu-vscode.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
+set -e
 bash scripts/os/ubuntu/profile-ubuntu-basic.sh
-sudo snap install code --classic
+if ! command -v code &>/dev/null; then
+    sudo snap install code --classic
+fi
+bash scripts/os/ubuntu/dep-vscode-settings.sh